ai-agent/pyproject.toml
[project]
name = "openemr-ai-agent"
version = "0.1.0"
description = "LangGraph AI agent for OpenEMR"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.132.0",
    "httpx>=0.27.0",
    "langchain-anthropic>=1.3.3",
    "langchain-core>=1.2.4",
    "langgraph>=0.2.70",
    "langsmith>=0.3.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.13.1",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "python-json-logger>=4.0.0",
    "pyyaml>=6.0.3",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "agentevals>=0.0.9",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-httpx>=0.36.0",
    "ruff>=0.15.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "unit: Fast unit tests (no Docker, no network)",
    "integration: Tests requiring Docker services (MySQL, OpenEMR API)",
    "slow: Tests that take more than a few seconds",
]
